+++ ionoscloud/list_resources.go
@@ -16,5 +16,6 @@ func ListResources() []func() list.ListResource {
 	return []func() list.ListResource{
 		NewDatacenterListResource,
 		NewIPBlockListResource,
+		NewDNSZoneListResource,
 	}
 }